% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angeles}
\alias{angeles}
\title{Weather in Los Angeles, CA}
\format{A data frame 7 rows and 3 variables:
   \describe{
   \item{date}{Date of weather observation}
   \item{TemperatureK}{Daily mean temperature in Kelvin}
   \item{DewpointK}{Daily mean dewpoint temperature in Kelvin}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
angeles
}
\description{
Daily values of mean temperature (Kelvin) and mean dew point
temperature (Kelvin) for the week of December 19, 2010, in Los Angeles, CA.
}
\keyword{datasets}

