% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/weathermetrics_data.R
=======
% Please edit documentation in R/data.R
>>>>>>> fb5490ec40e134247c15438f3f6f1c299693054e
\docType{data}
\name{beijing}
\alias{beijing}
\title{Weather in Beijing, China}
\format{A data frame with 7 rows and 3 variables:
     \describe{
     \item{date}{Date of weather observation}
     \item{TemperatureF}{Daily mean temperature in Fahrenheit}
     \item{MPH}{Daily mean wind speed in miles per hour}
     \item{mps}{Daily mean wind speed in meters per second}
     \item{ftps}{Daily mean wind speed in feet per second}
     \item{kmph}{Daily mean wind speed in kilometers per hour}
     }}
\source{
\href{http://www.wunderground.com}{Weather Underground}
}
\usage{
beijing
}
\description{
A dataset containing daily values of mean temperature (Fahrenheit) and mean
wind speed (in miles per hour, meters per second, feet per secons, and
kilometers per hour) for the week of January 10, 2015, in Beijing, China.
}
\keyword{datasets}

