% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breck}
\alias{breck}
\title{Precipitation in Breckenridge, CO}
\format{A data frame 7 rows and 2 variables:
   \describe{
   \item{date}{Date of weather observation}
   \item{Precip.in}{Daily precipitation in inches}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
breck
}
\description{
Daily values of precipitation (inches) for the week of June 28, 2015, in
Breckenridge, CO.
}
\keyword{datasets}

