% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_speed_conversion.R
\name{convert_wind_speed}
\alias{convert_wind_speed}
\title{Convert between standard units of measure for wind speed}
\usage{
convert_wind_speed(df, unit, to.knots = TRUE, round = TRUE)
}
\arguments{
\item{df}{Numeric vector of wind speeds}

\item{unit}{Character specifying the speed unit to convert from to knots, or
     convert to from knots.
Possible values are:
  \itemize{
    \item \code{mph}: Miles per hour
    \item \code{mps}: Meters per second
    \item \code{ftps}: Feet per second
    \item \code{kmph}: Kilometers per hour
  }}

\item{to.knots}{Variable indicating whether values should be converted to, or
from knots. Possible values are \code{TRUE} or \code{FALSE}. A value of
\code{TRUE} indicates values designated by \code{df} should be converted
to knots from the speed unit identified by \code{unit}. A value of
\code{FALSE} indicates values designated by \code{df} should be converted
from knots to the speed unit identified by \code{unit}.}

\item{round}{Variable indicating whether or not to round results. Possible
values are \code{TRUE} or \code{FALSE}. Default value is \code{TRUE}.
Values are rounded to 1 decimal place.}
}
\value{
A numeric vector of speeds (in knots, mph, mps, ftps, or kmph)
}
\description{
\code{convert_wind_speed} creates a numeric vector of speed in a specified unit of
   speed measurement from a numeric vector of speed in knots or a specified
   unit including miles per hour, meters per second, feet per second, or
   kilometers per hour.
}
\details{
The numeric vector specified for df must be in the unit knots, or
  as a character specified by \code{unit}.
}
\note{
Equations are from the source code for the National Oceanic and
   and Atmospheric Administration's
   \href{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}{online wind speed
   converter}
}
\examples{
data(beijing)
beijing$knots <- convert_wind_speed(beijing$kmph,
                     unit = "kmph",
                     to.knots = TRUE,
                     round = TRUE)
beijing
data(foco)
foco$mph <- convert_wind_speed(foco$knots,
                      unit = "mph",
                      to.knots = FALSE,
                      round = TRUE)
foco
}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Joshua Ferreri \email{joshua.m.ferreri@gmail.com}
}
\references{
\url{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}
}

