% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/weathermetrics_data.R
=======
% Please edit documentation in R/data.R
>>>>>>> fb5490ec40e134247c15438f3f6f1c299693054e
\docType{data}
\name{foco}
\alias{foco}
\title{Weather in Fort Collins, CO}
\format{A data frame with 7 rows and 3 variables:
     \describe{
     \item{date}{Date of weather observation}
     \item{TemperatureF}{Daily mean temperature in Fahrenheit}
     \item{knots}{Daily mean wind speed in knots}
     }}
\source{
\href{http://www.wunderground.com}{Weather Underground}
}
\usage{
foco
}
\description{
A dataset containing daily values of mean temperature (Fahrenheit) and mean
wind speed (in knots) for the week of October 11, 2015, in Fort
Collins, CO.
}
\keyword{datasets}

