% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainmeasure_conversion.R
\name{inches_to_metric}
\alias{inches_to_metric}
\title{Convert from inches to standard metric units of measure for precipitation}
\usage{
<<<<<<< HEAD
inches_to_metric(inches, unit, round = TRUE)
=======
inches_to_metric(inches, unit, round = 2)
>>>>>>> 46bb9ea8147fe862dd29d672b8368ccfe0933e6a
}
\arguments{
\item{inches}{Numeric vector of precipitation (in inches)}

\item{unit}{Character specifying the metric precipitation unit besides inches.
Possible values are:
  \itemize{
    \item \code{mm}: Millimeters
    \item \code{cm}: Centimeters
  }}

<<<<<<< HEAD
\item{round}{Variable indicating whether or not to round results.
Possible values are \code{TRUE} or \code{FALSE}. Default value is
\code{TRUE}. If \code{TRUE}, then the value will be rounded 2 spaces.}
=======
\item{round}{An integer indicating the number of decimal places to
round converted value.}
>>>>>>> 46bb9ea8147fe862dd29d672b8368ccfe0933e6a
}
\value{
A numeric vector of precipitation (in specified metric unit)
}
\description{
\code{inches_to_metric} creates a numeric vector of precipitation in
  common metric units (millimeters or centimeters) from a numeric vector of
  precipitation in inches.
}
\details{
The numeric vector specified for unit must be in the units provided
   as a character specified by unit.
}
\note{
Equations are from the source code for the National Weather Service
   Forecast Office \url{http://www.srh.noaa.gov/ama/?n=conversions}
}
\examples{
data(breck)
breck$Precip.mm <- inches_to_metric(breck$Precip.in,
                                    unit = "mm",
<<<<<<< HEAD
                                    round = TRUE)
=======
                                    round = 2)
>>>>>>> 46bb9ea8147fe862dd29d672b8368ccfe0933e6a
breck

}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Joshua Ferreri \email{joshua.m.ferreri@gmail.com}
}
\references{
\url{http://www.srh.noaa.gov/ama/?n=conversions}
}
\seealso{
\code{\link{metric_to_inches}}
}

