% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_conversions.R
\name{kelvin.to.fahrenheit}
\alias{kelvin.to.fahrenheit}
\title{Convert from Kelvin to Fahrenheit.}
\usage{
kelvin.to.fahrenheit(T.kelvin, round = 2)
}
\arguments{
\item{T.kelvin}{Numeric vector of temperatures in Kelvin.}

\item{round}{An integer indicating the number of decimal places to
round the converted value.}
}
\value{
A numeric vector of temperature values in Fahrenheit.
}
\description{
\code{kelvin.to.fahrenheit} creates a numeric vector of temperatures in
   Fahrenheit from a numeric vector of temperatures in Kelvin.
}
\note{
Equations are from the source code for the National Oceanic and
     Atmospheric Association's
    \href{http://www.srh.noaa.gov/epz/?n=wxcalc_tempconvert}{online
    temperature converter}.
}
\examples{
# Convert from Kelvin to Fahrenheit.
data(angeles)
angeles$TemperatureF <- kelvin.to.fahrenheit(angeles$TemperatureK)
angeles

}
\author{
Joshua Ferreri \email{joshua.m.ferreri@gmail.com},
Brooke Anderson \email{brooke.anderson@colostate.edu}
}
\seealso{
\code{\link{fahrenheit.to.kelvin}}
}

