% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{knots.to.speed}
\alias{knots.to.speed}
\title{Convert from knots to standard units of wind speed}
\usage{
\method{knots}{to.speed}(knots, round = 1, unit)
}
\arguments{
\item{knots}{Numeric vector of speeds in knots}

\item{round}{Integer indicating the number of decimal places to
round converted value. Default value is 1.}

\item{unit}{Character specifying the speed unit to convert to from knots.
Possible values are:
  \itemize{
    \item \code{mph}: Miles per hour
    \item \code{mps}: Meters per second
    \item \code{ftps}: Feet per second
    \item \code{kmph}: Kilometers per hour
  }}
}
\value{
A numeric vector of speeds (in the specified unit)
}
\description{
\code{knots.to.speed} creates a numeric vector of speed in a specified unit
  from a numeric vector of speed in knots.
}
\details{
Output will be in the speed units specified by unit.
}
\note{
Equations are from the source code for the National Oceanic and
  and Atmospheric Administration's
  \href{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}{online wind speed
  converter}
}
\examples{

}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Joshua Ferreri \email{joshua.m.ferreri@gmail.com}
}
\references{
\url{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}
}
\seealso{
\code{\link{speed.to.knots}}
}

