% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loveland}
\alias{loveland}
\title{Precipitation in Loveland, CO}
\format{A data frame 7 rows and 2 variables:
   \describe{
   \item{date}{Date of weather observation}
   \item{Precip.mm}{Daily precipitation in millimeters}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
loveland
}
\description{
Daily values of precipitation (millimeters) for the week of September 8,
2013, in Loveland, CO.
}
\keyword{datasets}
