\name{lyon}
\alias{lyon}
\docType{data}
\title{Weather in Lyon, France}
\description{
Daily values of mean temperature (Celsius) and mean dew point temperature (Celsius) for the week of June 18, 2000, in Lyon, France.
}
\usage{data(lyon)}
\format{
  A data frame with 7 observations on the following 3 variables.
  \describe{
    \item{\code{Date}}{a Date}
    \item{\code{TemperatureC}}{a numeric vector}
    \item{\code{DewpointC}}{a numeric vector}
  }
}
\source{
Weather Underground: http://www.wunderground.com/
}
\keyword{datasets}
