% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainmeasure_conversion.R
\name{metric.to.inches}
\alias{metric.to.inches}
\title{Convert between standard metric units of measure for precipitation to inches}
\usage{
metric.to.inches(metric, unit.from, round.out = TRUE, round = 2)
}
\arguments{
\item{metric}{Numeric vector of precipitation (in millimeters or centimeters)}

\item{unit.from}{Character specifying the metric precipitation unit convert
from to inches.
Possible values are:
  \itemize{
    \item \code{mm}: Millimeters
    \item \code{cm}: Centimeters
  }}

\item{round.out}{Variable indicating whether or not to round results.
Possible values are TRUE or FALSE. Default value is TRUE.}

\item{round}{Integer indicating the number of decimal places to
round converted value. Default value is 1.}
}
\value{
A numeric vector of precipitation in inches.
}
\description{
\code{metric.to.inches} creates a numeric vector of precipitation measures in
  inches from a numeric vector of precipitation in common metric units
  (millimeters or centimeters).
}
\details{
The numeric vector specified for unit.from must be in the units
   provided as a character specified by unit.from.
}
\note{
Equations are from the source code for the National Weather Service
   Forecast Office \href{http://www.srh.noaa.gov/ama/?n=conversions}
}
\examples{
data(loveland)
loveland$Precip.in <- metric.to.inches(loveland$Precip.mm,
                                       unit.from = "mm",
                                       round.out = TRUE,
                                       round = 2)
loveland

}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Joshua Ferreri \email{joshua.m.ferreri@gmail.com}
}
\references{
\url{http://www.srh.noaa.gov/ama/?n=conversions}
}
\seealso{
\code{\link{inches.to.metric}}
}

