\name{newhaven}
\alias{newhaven}
\docType{data}
\title{Weather in New Haven, CT}
\description{
Daily values of mean temperature (Fahrenheit) and mean relative humidity (\%) for the week of October 19, 2008, in New Haven, CT.
}
\usage{data(newhaven)}
\format{
  A data frame with 7 observations on the following 3 variables.
  \describe{
    \item{\code{Date}}{a Date}
    \item{\code{TemperatureF}}{a numeric vector}
    \item{\code{Relative.Humidity}}{a numeric vector}
  }
}
\source{
Weather Underground: http://www.wunderground.com/
}
\keyword{datasets}
