\name{norfolk}
\alias{norfolk}
\docType{data}
\title{Weather in Norfolk, VA}
\description{
Daily values of mean temperature (Fahrenheit) and mean dew point temperature (Fahrenheit) for the week of March 12, 2006, in Norfolk, VA.
}
\usage{data(norfolk)}
\format{
  A data frame with 7 observations on the following 3 variables.
  \describe{
    \item{\code{Date}}{a Date}
    \item{\code{TemperatureF}}{a numeric vector}
    \item{\code{DewpointF}}{a numeric vector}
  }
}
\source{
Weather Underground: http://www.wunderground.com/
}
\keyword{datasets}
