% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathermetrics_data.R
\docType{data}
\name{puravida}
\alias{puravida}
\title{Weather in San José, Costa Rica}
\format{A data frame 7 rows and 3 variables:
   \describe{
   \item{date}{Date of weather observation}
   \item{TemperatureF}{Daily mean temperature in Fahrenheit}
   \item{mph}{Daily mean wind speed in miles per hour}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
puravida
}
\description{
Daily values of mean temperature (Fahrenheit) and mean wind speed (miles per
hour) for the week of August 02, 2015, in San José, Costa Rica.
}
\keyword{datasets}

