% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{speed.to.knots}
\alias{speed.to.knots}
\title{Convert between standard units of measure for wind speed}
\usage{
speed.to.knots(x, round = 1, unit)
}
\arguments{
\item{x}{Numeric vector of wind speeds, in units specified by \code{unit}}

\item{round}{Integer indicating the number of decimal places to
round converted value. Default value is 1.}

\item{unit}{Character string specifying the speed unit for x. Possible values
are:
\itemize{
  \item \code{mph}: Miles per hour
  \item \code{mps}: Meters per second
  \item \code{ftps}: Feet per second
  \item \code{kmph}: Kilometers per hour
}}
}
\value{
A numeric vector of speeds (in knots)
}
\description{
\code{speed.to.knots} creates a numeric vector of speed in knots from a
  numeric vector of speed in the specified unit.
}
\details{
The numeric vector specified for x must be in the units provided as a
  character specified by unit.
}
\note{
Equations are from the source code for the National Oceanic and
   and Atmospheric Administration's
   \href{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}{online wind speed
   converter}
}
\examples{

}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Joshua Ferreri \email{joshua.m.ferreri@gmail.com}
}
\references{
\url{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}
}
\seealso{
\code{\link{knots.to.speed}}
}

