\name{suffolk}
\alias{suffolk}
\docType{data}
\title{Weather in Suffolk, VA}
\description{
Daily values of mean temperature (Fahrenheit) and mean relative humidity (\%) for the week of July 12, 1998, in Suffolk, VA.
}
\usage{data(suffolk)}
\format{
  A data frame with 7 observations on the following 3 variables.
  \describe{
    \item{\code{Date}}{a Date}
    \item{\code{TemperatureF}}{a numeric vector}
    \item{\code{Relative.Humidity}}{a numeric vector}
  }
}
\source{
Weather Underground: http://www.wunderground.com/
}
\keyword{datasets}
