% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weathermetrics.R
\docType{package}
\name{weathermetrics}
\alias{weathermetrics}
\alias{weathermetrics-package}
\title{weathermetrics: Functions to convert between weather metrics}
\description{
The weathermetics package provides functions to convert between Celsius
    and Fahrenheit, to convert between dew point temperature and relative
    humidity, and to calculate heat index.
}
\section{weathermetrics functions}{

This package includes the following functions to calculate vectors of
   weather metrics: 'celsius.to.fahrenheit', 'fahrenheit.to.celsius',
   'dewpoint.to.humidity', 'humidity.to.dewpoint', and 'heat.index'.
}

