% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainmeasure_conversion.R
\name{inches_to_metric}
\alias{inches_to_metric}
\title{Convert from inches to standard metric units of measure for precipitation}
\usage{
inches_to_metric(inches, unit, round.out = TRUE, round = 2)
}
\arguments{
\item{inches}{Numeric vector of precipitation (in inches)}

\item{unit}{Character specifying the metric precipitation unit to convert to
from inches.
Possible values are:
  \itemize{
    \item \code{mm}: Millimeters
    \item \code{cm}: Centimeters
  }}

\item{round.out}{Variable indicating whether or not to round results.
Possible values are TRUE or FALSE. Default value is TRUE.}

\item{round}{Integer indicating the number of decimal places to
round converted value. Default value is 1.}
}
\value{
A numeric vector of precipitation (in specified metric unit)
}
\description{
\code{inches_to_metric} creates a numeric vector of precipitation in
  common metric units (millimeters or centimeters) from a numeric vector of
  precipitation in inches.
}
\details{
The numeric vector specified for unit must be in the units provided
   as a character specified by unit.
}
\note{
Equations are from the source code for the National Weather Service
   Forecast Office \url{http://www.srh.noaa.gov/ama/?n=conversions}
}
\examples{
data(breck)
breck$Precip.mm <- inches_to_metric(breck$Precip.in,
                                    unit = "mm",
                                    round.out = TRUE,
                                    round = 2)
breck

}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Joshua Ferreri \email{joshua.m.ferreri@gmail.com}
}
\references{
\url{http://www.srh.noaa.gov/ama/?n=conversions}
}
\seealso{
\code{\link{metric_to_inches}}
}

