% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_conversions.R
\name{knots_to_speed}
\alias{knots_to_speed}
\title{Convert from knots to standard units of wind speed}
\usage{
knots_to_speed(knots, unit, round.out = TRUE, round = 1)
}
\arguments{
\item{knots}{Numeric vector of speeds in knots}

\item{unit}{Character specifying the speed unit to convert to from knots.
Possible values are:
  \itemize{
    \item \code{mph}: Miles per hour
    \item \code{mps}: Meters per second
    \item \code{ftps}: Feet per second
    \item \code{kmph}: Kilometers per hour
  }}

\item{round.out}{Variable with the indicating whether or not to round results.
Possible values are TRUE or FALSE. Default value is TRUE.}

\item{round}{Integer indicating the number of decimal places to
round converted value. Default value is 1.}
}
\value{
A numeric vector of speeds (in the specified unit)
}
\description{
\code{knots_to_speed} creates a numeric vector of speed, in units
   specified by \code{unit},
  from a numeric vector of speed in knots.
}
\details{
Output will be in the speed units specified by unit.
}
\note{
Equations are from the source code for the National Oceanic and
  and Atmospheric Administration's
  \href{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}{online wind speed
  converter}
}
\examples{
data(foco)
foco$mph <- knots_to_speed(foco$knots, "mph", round.out = TRUE)
foco$mps <- knots_to_speed(foco$knots, "mps", round.out = TRUE)
foco$ftps <- knots_to_speed(foco$knots, "ftps", round.out = TRUE)
foco$kmph <- knots_to_speed(foco$knots, "kmph", round.out = TRUE)
foco
}
\author{
Brooke Anderson \email{brooke.anderson@colostate.edu},
Joshua Ferreri \email{joshua.m.ferreri@gmail.com}
}
\references{
\url{http://www.srh.noaa.gov/epz/?n=wxcalc_windconvert}
}
\seealso{
\code{\link{speed_to_knots}}
}

