% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyon.R
\docType{data}
\name{lyon}
\alias{lyon}
\title{Weather in Lyon, France}
\format{A data frame with columns:
   \describe{
   \item{Date}{Date of weather observation}
   \item{TemperatureC}{Daily mean temperature in Celsius}
   \item{DewpointC}{Daily mean dewpoint temperature in Celsius}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
lyon
}
\description{
Daily values of mean temperature (Celsius) and mean dew point temperature
(Celsius) for the week of June 18, 2000, in Lyon, France.
}
\keyword{datasets}

