% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newhaven}
\alias{newhaven}
\title{Weather in New Haven, CT}
\format{A data frame with columns:
   \describe{
   \item{Date}{Date of weather observation}
   \item{TemperatureF}{Daily mean temperature in Fahrenheit}
   \item{Relative.Humidity}{Daily relative humidity in \%}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
newhaven
}
\description{
Daily values of mean temperature (Fahrenheit) and mean relative humidity
(\%) for the week of October 19, 2008, in New Haven, CT.
}
\keyword{datasets}

