% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norfolk.R
\docType{data}
\name{norfolk}
\alias{norfolk}
\title{Weather in Norfolk, VA}
\format{A data frame with columns:
   \describe{
   \item{Date}{Date of weather observation}
   \item{TemperatureF}{Daily mean temperature in Fahrenheit}
   \item{DewpointF}{Daily mean dewpoint temperature in Fahrenheit}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
norfolk
}
\description{
Daily values of mean temperature (Fahrenheit) and mean dew point
temperature (Fahrenheit) for the week of March 12, 2006, in Norfolk, VA.
}
\keyword{datasets}

