% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffolk.R
\docType{data}
\name{suffolk}
\alias{suffolk}
\title{Weather in Suffolk, VA}
\format{A data frame with columns:
   \describe{
   \item{Date}{Date of weather observation}
   \item{TemperatureF}{Daily mean temperature in Fahrenheit}
   \item{Relative.Humidity}{Daily relative humidity in \%}
   }}
\source{
\href{http://www.wunderground.com/}{Weather Underground}
}
\usage{
suffolk
}
\description{
Daily values of mean temperature (Fahrenheit) and mean relative humidity
(\%) for the week of July 12, 1998, in Suffolk, VA.
}
\keyword{datasets}

